MORE SPEEEED

+++ GP1_Raytracer-main/source/Vector3.cpp
@@ -19,7 +19,7 @@ namespace dae {
 
 	float Vector3::Magnitude() const
 	{
-		return std::sqrtf(x * x + y * y + z * z);
+		return std::sqrt(x * x + y * y + z * z);
 	}
 
 	float Vector3::SqrMagnitude() const
